Fix segfault when exiting in ruby 1.9
In ruby 1.8, the :define_finalizer method would call all of the procs *before* the object was actually destroyed while in 1.9 they are called *after* they are destroyed.

The object we're calling a finalizer on is a global variable anyway and will only be destroyed at the end of the process, so just send an end callback instead. It was confirmed that these are run before the actual object is destroyed and this prevents that form of a segfault.